% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_dir}
\alias{explore_dir}
\title{Make an interactive shiny app to explore the directory}
\usage{
explore_dir(path, files = NULL)
}
\arguments{
\item{path}{path to directory to visualize}

\item{files}{keep only these files in nested list (default to NULL)}
}

