% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_dir}
\alias{nest_dir}
\title{nest_dir}
\usage{
nest_dir(path, value = bname)
}
\arguments{
\item{path}{Base directory to start nesting from}

\item{value}{function to assign a value for each node (default `bname`)}
}
\value{
nested list
}
\description{
Make a nested list out of a directory
Note: there might be problems with document name length on windows.
tools::file_ext() was the quickest check for directory I could think of,
but there is probably something better (pattern "/$" or similar?)
}

