% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{remove_nulls}
\alias{remove_nulls}
\title{remove_nulls}
\usage{
remove_nulls(nest)
}
\arguments{
\item{nest}{nested list}
}
\value{
nested list with all NULLs removed (and all branches containing only NULLs)
}

