% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{trim_nest}
\alias{trim_nest}
\title{Keep only specific files in nested directories}
\usage{
trim_nest(lst, files)
}
\arguments{
\item{lst}{nested list}

\item{files}{values to keep in nested list}
}
\value{
nested list only containing files (and branches with those files)
}
\description{
Searches through nested directory list and removes entries not
containing specified files.  If an entire branch of the nested list
contains no specified files, it will be removed entirely.
}

