% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{bname}
\alias{bname}
\title{bname}
\usage{
bname(ff)
}
\arguments{
\item{ff}{Filename or empty list (result from nest_dir)}
}
\value{
basename of file or an empty string
}
\description{
Helper function to return a value for nest_dir
(empty strings for empty lists)
}

